fix: PerformanceSearchRepositoryTest이 실행되지 않는 오류 수정
