+++ src/test/kotlin/com/flab/ticketing/performance/integration/PerformanceIntegrationTest.kt
@@ -347,7 +347,7 @@ class PerformanceIntegrationTest : IntegrationTest() {
 
             savePerformance(performances)
             `when`("사용자가 5개의 공연 정보를 조회할 시") {
-                val uri = "/api/v2/performances"
+                val uri = "/api/performances"
                 val limit = 5
 
                 val mvcResult = mockMvc.perform(
